REF: collect .align tests (#33019)
