+++ pandas/tests/series/indexing/test_alter_index.py
@@ -8,162 +8,6 @@ from pandas import Categorical, Series, date_range, isna
 import pandas._testing as tm
 
 
-@pytest.mark.parametrize(
-    "first_slice,second_slice",
-    [
-        [[2, None], [None, -5]],
-        [[None, 0], [None, -5]],
-        [[None, -5], [None, 0]],
-        [[None, 0], [None, 0]],
-    ],
-)
-@pytest.mark.parametrize("fill", [None, -1])
-def test_align(datetime_series, first_slice, second_slice, join_type, fill):
-    a = datetime_series[slice(*first_slice)]
-    b = datetime_series[slice(*second_slice)]
-
-    aa, ab = a.align(b, join=join_type, fill_value=fill)
-
-    join_index = a.index.join(b.index, how=join_type)
-    if fill is not None:
-        diff_a = aa.index.difference(join_index)
-        diff_b = ab.index.difference(join_index)
-        if len(diff_a) > 0:
-            assert (aa.reindex(diff_a) == fill).all()
-        if len(diff_b) > 0:
-            assert (ab.reindex(diff_b) == fill).all()
-
-    ea = a.reindex(join_index)
-    eb = b.reindex(join_index)
-
-    if fill is not None:
-        ea = ea.fillna(fill)
-        eb = eb.fillna(fill)
-
-    tm.assert_series_equal(aa, ea)
-    tm.assert_series_equal(ab, eb)
-    assert aa.name == "ts"
-    assert ea.name == "ts"
-    assert ab.name == "ts"
-    assert eb.name == "ts"
-
-
-@pytest.mark.parametrize(
-    "first_slice,second_slice",
-    [
-        [[2, None], [None, -5]],
-        [[None, 0], [None, -5]],
-        [[None, -5], [None, 0]],
-        [[None, 0], [None, 0]],
-    ],
-)
-@pytest.mark.parametrize("method", ["pad", "bfill"])
-@pytest.mark.parametrize("limit", [None, 1])
-def test_align_fill_method(
-    datetime_series, first_slice, second_slice, join_type, method, limit
-):
-    a = datetime_series[slice(*first_slice)]
-    b = datetime_series[slice(*second_slice)]
-
-    aa, ab = a.align(b, join=join_type, method=method, limit=limit)
-
-    join_index = a.index.join(b.index, how=join_type)
-    ea = a.reindex(join_index)
-    eb = b.reindex(join_index)
-
-    ea = ea.fillna(method=method, limit=limit)
-    eb = eb.fillna(method=method, limit=limit)
-
-    tm.assert_series_equal(aa, ea)
-    tm.assert_series_equal(ab, eb)
-
-
-def test_align_nocopy(datetime_series):
-    b = datetime_series[:5].copy()
-
-    # do copy
-    a = datetime_series.copy()
-    ra, _ = a.align(b, join="left")
-    ra[:5] = 5
-    assert not (a[:5] == 5).any()
-
-    # do not copy
-    a = datetime_series.copy()
-    ra, _ = a.align(b, join="left", copy=False)
-    ra[:5] = 5
-    assert (a[:5] == 5).all()
-
-    # do copy
-    a = datetime_series.copy()
-    b = datetime_series[:5].copy()
-    _, rb = a.align(b, join="right")
-    rb[:3] = 5
-    assert not (b[:3] == 5).any()
-
-    # do not copy
-    a = datetime_series.copy()
-    b = datetime_series[:5].copy()
-    _, rb = a.align(b, join="right", copy=False)
-    rb[:2] = 5
-    assert (b[:2] == 5).all()
-
-
-def test_align_same_index(datetime_series):
-    a, b = datetime_series.align(datetime_series, copy=False)
-    assert a.index is datetime_series.index
-    assert b.index is datetime_series.index
-
-    a, b = datetime_series.align(datetime_series, copy=True)
-    assert a.index is not datetime_series.index
-    assert b.index is not datetime_series.index
-
-
-def test_align_multiindex():
-    # GH 10665
-
-    midx = pd.MultiIndex.from_product(
-        [range(2), range(3), range(2)], names=("a", "b", "c")
-    )
-    idx = pd.Index(range(2), name="b")
-    s1 = pd.Series(np.arange(12, dtype="int64"), index=midx)
-    s2 = pd.Series(np.arange(2, dtype="int64"), index=idx)
-
-    # these must be the same results (but flipped)
-    res1l, res1r = s1.align(s2, join="left")
-    res2l, res2r = s2.align(s1, join="right")
-
-    expl = s1
-    tm.assert_series_equal(expl, res1l)
-    tm.assert_series_equal(expl, res2r)
-    expr = pd.Series([0, 0, 1, 1, np.nan, np.nan] * 2, index=midx)
-    tm.assert_series_equal(expr, res1r)
-    tm.assert_series_equal(expr, res2l)
-
-    res1l, res1r = s1.align(s2, join="right")
-    res2l, res2r = s2.align(s1, join="left")
-
-    exp_idx = pd.MultiIndex.from_product(
-        [range(2), range(2), range(2)], names=("a", "b", "c")
-    )
-    expl = pd.Series([0, 1, 2, 3, 6, 7, 8, 9], index=exp_idx)
-    tm.assert_series_equal(expl, res1l)
-    tm.assert_series_equal(expl, res2r)
-    expr = pd.Series([0, 0, 1, 1] * 2, index=exp_idx)
-    tm.assert_series_equal(expr, res1r)
-    tm.assert_series_equal(expr, res2l)
-
-
-@pytest.mark.parametrize("method", ["backfill", "bfill", "pad", "ffill", None])
-def test_align_method(method):
-    # GH31788
-    ser = pd.Series(range(3), index=range(3))
-    df = pd.DataFrame(0.0, index=range(3), columns=range(3))
-
-    result_ser, result_df = ser.align(df, method=method)
-    tm.assert_series_equal(result_ser, ser)
-    tm.assert_frame_equal(result_df, df)
-
-
 def test_reindex(datetime_series, string_series):
     identity = string_series.reindex(string_series.index)
